organize class into commands, exceptions and tasks packages
