+++ src/main/java/nyx/Parser.java
@@ -1,3 +1,16 @@
+package nyx;
+
+import nyx.commands.ByeCommand;
+import nyx.commands.Command;
+import nyx.commands.DeadlineCommand;
+import nyx.commands.DeleteCommand;
+import nyx.commands.EventCommand;
+import nyx.commands.ListCommand;
+import nyx.commands.MarkCommand;
+import nyx.commands.TodoCommand;
+import nyx.commands.UnknownCommand;
+import nyx.commands.UnmarkCommand;
+
 public class Parser {
 
     private enum CommandType {
