+++ src/main/java/nyx/Storage.java
@@ -1,3 +1,5 @@
+package nyx;
+
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.FileWriter;
@@ -8,6 +10,11 @@ import java.nio.file.Paths;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
+
+import nyx.tasks.DeadlineTask;
+import nyx.tasks.EventTask;
+import nyx.tasks.Task;
+import nyx.tasks.TodoTask;
 
 public class Storage {
     private static final Path FILE_PATH = Paths.get("data", "tasks.txt");
