+++ src/main/java/nyx/TaskList.java
@@ -1,7 +1,11 @@
+package nyx;
+
 import java.util.ArrayList;
 
+import nyx.tasks.Task;
+
 public class TaskList {
-    private ArrayList<Task> tasks;
+    private final ArrayList<Task> tasks;
 
     public TaskList(ArrayList<Task> tasks) {
         this.tasks = tasks;
