+++ src/main/java/nyx/commands/MarkCommand.java
@@ -1,5 +1,10 @@
-import exceptions.InvalidUsageException;
-import exceptions.NyxException;
+package nyx.commands;
+
+import nyx.exceptions.InvalidUsageException;
+import nyx.exceptions.NyxException;
+import nyx.Storage;
+import nyx.TaskList;
+import nyx.Ui;
 
 public class MarkCommand extends Command {
     private final String command;
