+++ src/main/java/nyx/commands/TodoCommand.java
@@ -1,7 +1,13 @@
-import exceptions.InvalidUsageException;
-import exceptions.NyxException;
+package nyx.commands;
 
-public class TodoCommand extends Command{
+import nyx.exceptions.InvalidUsageException;
+import nyx.exceptions.NyxException;
+import nyx.Storage;
+import nyx.TaskList;
+import nyx.Ui;
+import nyx.tasks.TodoTask;
+
+public class TodoCommand extends Command {
 
     private final String command;
