+++ src/main/java/nyx/commands/UnknownCommand.java
@@ -1,5 +1,10 @@
-import exceptions.NyxException;
-import exceptions.UnknownCommandException;
+package nyx.commands;
+
+import nyx.exceptions.NyxException;
+import nyx.exceptions.UnknownCommandException;
+import nyx.Storage;
+import nyx.TaskList;
+import nyx.Ui;
 
 public class UnknownCommand extends Command {
     public void execute(TaskList taskList, Storage storage, Ui ui) throws NyxException {
