+++ src/main/java/nyx/exceptions/InvalidUsageException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package nyx.exceptions;
 
 public class InvalidUsageException extends NyxException {
     public InvalidUsageException(String message) {
