+++ src/main/java/nyx/exceptions/NyxException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package nyx.exceptions;
 
 public class NyxException extends Exception {
     public NyxException(String message) {
