+++ src/main/java/nyx/exceptions/UnknownCommandException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package nyx.exceptions;
 
 public class UnknownCommandException extends NyxException {
     public UnknownCommandException(String message) {
