+++ src/main/java/nyx/tasks/EventTask.java
@@ -1,3 +1,5 @@
+package nyx.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
