+++ src/main/java/nyx/tasks/TodoTask.java
@@ -1,3 +1,5 @@
+package nyx.tasks;
+
 public class TodoTask extends Task {
     public TodoTask(String name) {
         super(name);
